refactor(executor): decompose nested_loop.rs into focused modules
Split the 1637-line nested_loop.rs into logical modules:

- condition.rs: Join condition analysis and evaluation helpers (~90 lines)
- memory_guard.rs: Memory size limit checking (~30 lines)
- sequential.rs: Sequential nested loop implementations (~230 lines)
- parallel.rs: Parallel nested loop with work-stealing (~220 lines)
- mod.rs: Join type implementations (INNER, LEFT/RIGHT/FULL OUTER, CROSS, SEMI, ANTI)
- tests.rs: Comprehensive test suite including parallel tests

Benefits:
- Sequential and parallel implementations clearly separated
- Condition analysis reusable by other join algorithms
- Memory guard logic shared across join types
- Easier to optimize individual execution paths
- Main mod.rs reduced to ~400 lines

All 13 nested_loop tests pass.

Closes #4870
